backend/requirements.txt
sentence-transformers
faiss-cpu
pydantic
emoji
